% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4d-package.R
\docType{package}
\name{a4d-package}
\alias{a4d}
\alias{a4d-package}
\title{a4d: Process tracker data for the DataCross A4D project}
\description{
Functionality to process, clean and ingest tracker data (excel files).
}
\author{
\strong{Maintainer}: Michael Aydinbas \email{michael.aydinbas@gmail.com}

}
\keyword{internal}
