% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{add_product_categories}
\alias{add_product_categories}
\title{Inserts product categories if contained in a provided mapping}
\usage{
add_product_categories(inventory_data, product_category_mapping)
}
\arguments{
\item{inventory_data}{A product inventory dataframe containing a "product" column}

\item{product_category_mapping}{A dataframe mapping "product" to "category"}
}
\value{
A dataframe containing the inventory data with an additional product
category column.
}
\description{
Product categories are added in the sense of a left join,
meaning missing categories will result in NA values while products are never
dropped.
}
