% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{bmi_fix}
\alias{bmi_fix}
\title{Set bmi value to NA if height or weight is not available.}
\usage{
bmi_fix(df)
}
\arguments{
\item{df}{data frame with patient data.}
}
\value{
data frame with corrected bmi column.
}
\description{
Set bmi value to NA if height or weight is not available.
}
