% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_product_data.R
\name{calculate_most_frequent}
\alias{calculate_most_frequent}
\title{Calculate the Most Frequent Value in a Data Frame}
\usage{
calculate_most_frequent(df, group_column, value_column, new_column)
}
\arguments{
\item{df}{A data frame.}

\item{group_column}{The name of the column to group by.}

\item{value_column}{The name of the column to calculate the most frequent value from.}

\item{new_column}{The name of the new column to add to the data frame.}
}
\value{
A data frame with an additional column containing the most frequent value of the specified column for each group.
}
\description{
This function calculates the most frequent value of a specified column for each group in a data frame.
It then adds a new column to the data frame with these most frequent values.
}
\examples{
\dontrun{
library(dplyr)
data(mtcars)
result <- calculate_most_frequent(mtcars, "cyl", "gear", "most_frequent_gear")
}
}
