% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_A4D_directory.R
\name{change_A4D_directory}
\alias{change_A4D_directory}
\title{Changes the path to A4D data that is set as an environment variable}
\usage{
change_A4D_directory()
}
\value{
Changes the env variable to A4D data folder
}
\description{
Changes the path to A4D data that is set as an environment variable
}
