% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{check_entry_dates}
\alias{check_entry_dates}
\title{Check entry dates in product data}
\usage{
check_entry_dates(df, Sheet)
}
\arguments{
\item{df}{A dataframe containing product data.}

\item{Sheet}{The name of the sheet where the product data is located.}
}
\value{
This function does not return a value. It logs a warning message if there are any dates in 'product_entry_date' that don't match the month/year on the sheet.
}
\description{
This function checks if the entry dates for the balance match the month/year on the sheet. If any discrepancies are found, it logs a warning message with the number of mismatched dates and their corresponding 'product_entry_date' values.
}
