% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{check_negative_balance}
\alias{check_negative_balance}
\title{Check negative values in 'product_balance' column}
\usage{
check_negative_balance(df, Sheet)
}
\arguments{
\item{df}{A dataframe containing product data.}

\item{Sheet}{The name of the sheet where the product data is located.}
}
\value{
This function does not return a value. It logs a warning message if there are any negative values in the 'product_balance' column.
}
\description{
This function checks for negative values in the 'product_balance' column of a given dataframe. If any negative values are found, it logs a warning message with the number of negative values and their corresponding 'product_balance' values.
}
