% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_format_2_long_format.R
\name{check_patterns_in_column}
\alias{check_patterns_in_column}
\title{Check for Specific Patterns in a Dataframe Column}
\usage{
check_patterns_in_column(df, column_name)
}
\arguments{
\item{df}{A dataframe to check for the specific patterns.}

\item{column_name}{The name of the column to check for the patterns.}
}
\value{
A logical value indicating whether any cell in the specified column contains the patterns.
}
\description{
This function checks if any cell in a specific column of a dataframe contains a pattern with a hyphen followed by a digit, or a comma (which could be leading with a space). It returns TRUE if any cell in the column contains the pattern, and FALSE otherwise.
}
\examples{
data(mtcars)
check_patterns_in_column(mtcars, "mpg")

}
