% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_format_2_long_format.R
\name{check_wide_format_columns}
\alias{check_wide_format_columns}
\title{Check if dataframe contains required columns}
\usage{
check_wide_format_columns(df)
}
\arguments{
\item{df}{A dataframe to check for the presence of required columns.}
}
\value{
A logical value indicating whether the dataframe contains the required columns.
}
\description{
This function checks if a dataframe contains the columns 'Total Units Released' and 'Units Released per person'.
(for detection the Mandalay Children's Hospital 2020-2021 wide-format)
It returns TRUE if both columns are present, and FALSE otherwise.
}
\examples{
df <- data.frame("Total Units Released" = c(1, 2, 3), "Units Released per person" = c(4, 5, 6), "Other Column" = c(7, 8, 9))
check_wide_format_columns(df) # Returns: TRUE
}
