% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{clean_unitsreceived}
\alias{clean_unitsreceived}
\title{clean_unitsreceived}
\usage{
clean_unitsreceived(product_df)
}
\description{
Clean column "product_units_received" from character values and rows with values but no further information
}
