% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{compare_lists}
\alias{compare_lists}
\title{Compare two lists and return unmatched strings}
\usage{
compare_lists(list1, list2)
}
\arguments{
\item{list1}{The first list of strings to be compared.}

\item{list2}{The second list of strings to be compared.}
}
\value{
A vector of strings that are present in 'list1' but not in 'list2'.
}
\description{
This function compares two lists and returns the strings that are present in the first list but not in the second list.
}
