% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{compute_balance}
\alias{compute_balance}
\title{compute_balance}
\usage{
compute_balance(product_df, year)
}
\description{
Computes balance based on start balance value, units released, units returned, units received.-
}
