% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{compute_balance_cleanrows}
\alias{compute_balance_cleanrows}
\title{compute_balance_cleanrows}
\usage{
compute_balance_cleanrows(product_df)
}
\description{
Preparation function for compute_balance and compute_balance_status.

Remove rows without any meaningful information regarding units (release, return, received) or entry status
}
