% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{compute_balance_status}
\alias{compute_balance_status}
\title{compute_balance_status}
\usage{
compute_balance_status(product_df)
}
\description{
Computate balance status (start vs. change vs. end) per product
}
