% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{correct_decimal_sign}
\alias{correct_decimal_sign}
\title{Correct decimal sign.}
\usage{
correct_decimal_sign(x)
}
\arguments{
\item{x}{character representation of a number.}
}
\value{
character representation with a correct decimal sign.
}
\description{
Correct decimal sign.
}
