% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{count_na_rows}
\alias{count_na_rows}
\title{Count rows with missing patient's name next to the released units}
\usage{
count_na_rows(df, units_released_col, released_to_col)
}
\arguments{
\item{df}{A dataframe containing patient data.}

\item{units_released_col}{The name of the column that contains information about the units released.}

\item{released_to_col}{The name of the column that contains information about who the units were released to (patient's name).}
}
\value{
The number of rows where the patient's name is missing next to the released units.
}
\description{
This function checks if the patient's name is missing next to the released units in a given dataframe. It counts the number of such rows.
}
