% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_format_2_long_format.R
\name{create_new_rows}
\alias{create_new_rows}
\title{Create New Rows in a DataFrame}
\usage{
create_new_rows(df)
}
\arguments{
\item{df}{A dataframe containing the original data.}
}
\value{
A modified dataframe with additional rows.
}
\description{
This function creates new rows in a given dataframe based on specific conditions.
To convert wide-format area in the 2017 - 2020 data to long-format area.
}
\details{
The function checks if 'Total Units Released' and 'Units Released per person'
are not empty and not equal. If this condition is met, it adds new rows to the dataframe
based on values in columns between 'Released To' and 'Units Released per person'.
If an error occurs during processing, the original dataframe is returned.
}
\examples{
df <- data.frame(
    "Date" = c("2024-03-01", "2024-03-02"),
    "Total Units Released" = c(100, 200),
    "Units Released per person" = c(10, 20),
    "Released To (select from drop down list)" = c("A", "B")
)
new_df <- create_new_rows(df)
}
