% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_clinic_static_data.R
\name{create_table_clinic_static_data}
\alias{create_table_clinic_static_data}
\title{Create clinic static data table}
\usage{
create_table_clinic_static_data(output_root)
}
\arguments{
\item{output_root}{The root directory of the output folder.}
}
\description{
This function reads the clinic data from an Excel file, fills missing values, and exports the data as a parquet file.
}
