% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script3_create_table_patient_data_annual.R
\name{create_table_patient_data_annual}
\alias{create_table_patient_data_annual}
\title{Create CSV with annual patient data}
\usage{
create_table_patient_data_annual(patient_data_files, input_root, output_root)
}
\arguments{
\item{patient_data_files}{list of CSV files with cleaned patient data from step 2.}

\item{input_root}{root directory of the input CSV files.}

\item{output_root}{root directory of the output folder.}
}
\description{
Read in all cleaned patient data CSV and create a single data.frame.
Group this data by id and take the latest available data (latest year and month).
}
