% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_patient_data.R
\name{create_table_patient_data_monthly}
\alias{create_table_patient_data_monthly}
\title{Create CSV with monthly patient data}
\usage{
create_table_patient_data_monthly(all_patient_data, output_folder)
}
\arguments{
\item{all_patient_data}{list of data frames with cleaned patient data from step 2.}

\item{output_folder}{output directory.}
}
\description{
Read in all cleaned patient data CSV and create a single data.frame.
Only take dynamic columns.
}
