% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_patient_data_static.R
\name{create_table_patient_data_static}
\alias{create_table_patient_data_static}
\title{Create CSV with static patient data}
\usage{
create_table_patient_data_static(all_patient_data, output_folder)
}
\arguments{
\item{all_patient_data}{list of data frames with cleaned patient data from step 2.}

\item{output_folder}{output directory.}
}
\description{
Read in all cleaned patient data CSV and create a single data.frame.
Group this data by id and take the latest available data (latest year and month).
}
