% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_product_data.R
\name{create_table_product_data}
\alias{create_table_product_data}
\title{Create a Single CSV for Table Product Data}
\usage{
create_table_product_data(all_product_data, output_folder)
}
\arguments{
\item{all_product_data}{list of data frames with cleaned product data from step 2.}

\item{output_folder}{output irectory where the output .parquet file should be written.}
}
\value{
This function does not return a value. It writes the merged data to a new CSV file
(with reordered columns according to the list of fields) in the output_root directory.
}
\description{
This function reads all parquet files in a specified directory, merges them into a single data frame,
and writes the merged data to a new file with reordered and processed columns according to the list of fields.
If there are any columns that are present in some files but not others,
this function will add those missing columns to the data frames where they are missing and fill them with NA.
}
\examples{
\dontrun{
create_table_product_data(list(df1, df2, ...), "path/to/output/directory")
}
}
