% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{date_as_fivedigit_number_fix}
\alias{date_as_fivedigit_number_fix}
\title{WIP: Convert dates that are stored as 5-digit number from Excel to date format}
\usage{
date_as_fivedigit_number_fix(d)
}
\arguments{
\item{d}{}
}
\value{
Date in a character format.
}
\description{
WIP: Convert dates that are stored as 5-digit number from Excel to date format
}
\examples{
date_as_fivedigit_number_fix("44939") # res in "2023-01-13"

}
\seealso{
https://stackoverflow.com/a/50274507: depending on whether read in
file was created on Mac or Windows, the origin date might need to be adjusted.
}
