% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_main.R
\name{export_data}
\alias{export_data}
\title{Export data as CSV to a given destination.}
\usage{
export_data(data, filename, output_root, suffix)
}
\arguments{
\item{data}{Data frame to export as CSV.}

\item{filename}{Output file name.}

\item{output_root}{Root output directory.}

\item{suffix}{Suffix will be appended to the original file name (e.g. "patient_data").}
}
\description{
Export data as CSV to a given destination.
}
\examples{
\dontrun{
export_data(
    data = df_raw_product,
    filename = tracker_name,
    output_root = output_root,
    suffix = "_product_data"
)
}
}
