% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{extract_date_from_measurement}
\alias{extract_date_from_measurement}
\title{Split a column value into value and date.}
\usage{
extract_date_from_measurement(df, colname)
}
\arguments{
\item{df}{data frame with patient data.}

\item{colname}{column to split into value and date.}
}
\value{
data frame with the original column and a new colname_date column.
}
\description{
Assumes that the measurement entry is in the format "8.53 (28/8/2017)".
Uses regex to split this string into the value part and the date part,
for example in "8.53 " and "28/8/2017".
}
