% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{extract_first_raw_regimen}
\alias{extract_first_raw_regimen}
\title{Extract the part of a string before a comma}
\usage{
extract_first_raw_regimen(string)
}
\arguments{
\item{string}{The input string}
}
\value{
The extracted part.
}
\description{
Extract the part of a string before a comma
}
