% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_read_patient_data.R
\name{extract_patient_data}
\alias{extract_patient_data}
\title{Extract the patient data from a month sheet of a tracker file.}
\usage{
extract_patient_data(tracker_data, year)
}
\arguments{
\item{tracker_data}{data.frame holding the data from a month sheet.}

\item{year}{year of this tracker.}
}
\value{
data.frame with the patient data
}
\description{
Searches for first and last row that starts with country and clinic code
and extracts the data from this range as data.frame
}
