% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_read_patient_data.R
\name{extract_patient_data_header}
\alias{extract_patient_data_header}
\title{Extract patient data header names from the month sheet of a tracker file.}
\usage{
extract_patient_data_header(tracker_data, year)
}
\arguments{
\item{tracker_data}{data.frame holding the data from a month sheet.}

\item{year}{year of this tracker.}
}
\value{
vector with column names.
}
\description{
Search for the cell with "Patient ID" and extract that row as header.
For trackers from 2019 and newer the header spans two columns,
so date is added to the column name.
}
