% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{extract_product_data}
\alias{extract_product_data}
\title{extract_product_data}
\usage{
extract_product_data(monthly_tracker_df)
}
\description{
Reads product data from a monthly file based on extraction logic
}
