% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{extract_product_multiple}
\alias{extract_product_multiple}
\title{extract_product_multiple}
\usage{
extract_product_multiple(product_df)
}
\description{
Splits product name cells with multiple values into separate rows. Uses numeric product content in cell for columns units_received or units_released
}
