% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{extract_regimen}
\alias{extract_regimen}
\title{Maps Insulin regiment with allowed values.}
\usage{
extract_regimen(raw_input)
}
\arguments{
\item{raw_input}{The string that needs to be mapped}
}
\value{
The mapping result
}
\description{
The mapping is done using regular expressions.
}
