% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{extract_unit_capacity}
\alias{extract_unit_capacity}
\title{Extract Unit Capacity from a Specified Column}
\usage{
extract_unit_capacity(df, column_name)
}
\arguments{
\item{df}{A dataframe that contains the column to extract unit capacity from.}

\item{column_name}{The name of the column to extract unit capacity from.}
}
\value{
A dataframe with an additional column 'product_unit_capacity' that contains the extracted unit capacity.
}
\description{
This function extracts the unit capacity from a specified column in a dataframe.
It assumes that the unit capacity is represented by numbers immediately before 's or s within parentheses.
Non-numeric characters are removed and the extracted values are converted to numeric.
NA values are replaced with 1.
}
\examples{
\dontrun{
df <- data.frame(product = c("Product A (2s)", "Product B (3's)", "Product C"))
df <- extract_unit_capacity(df, "product")
}
}
