% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{extract_year_from_age}
\alias{extract_year_from_age}
\title{Extract year from age descriptions like "10y5m"}
\usage{
extract_year_from_age(age)
}
\arguments{
\item{age}{text representation of diagnosis age}
}
\value{
year of given age as character.
}
\description{
Example: "1y5m" -> "1"
}
