% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_age}
\alias{fix_age}
\title{Fix age column.}
\usage{
fix_age(age, dob, tracker_year, tracker_month, id)
}
\arguments{
\item{age}{patient age.}

\item{dob}{patient date of birth.}

\item{tracker_year}{year the tracker was filled out.}

\item{tracker_month}{month the tracker was filled out.}

\item{id}{patient id.}
}
\value{
fixed age.
}
\description{
Based on date of birth calculate the current age of the patient.
If this age is different than the given patient age, we take the calculated one.
}
