% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_bmi}
\alias{fix_bmi}
\title{Only accept value for bmi if there is also a weight and height.}
\usage{
fix_bmi(bmi, weight, height, id)
}
\arguments{
\item{bmi}{numeric bmi value.}

\item{weight}{numeric weight value.}

\item{height}{numeric height value.}

\item{id}{patient id.}
}
\value{
corrected bmi value
}
\description{
Set bmi to error val if either weight or height is NA.
}
