% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_digit_date}
\alias{fix_digit_date}
\title{WIP: Convert dates that are stored as 5-digit number from Excel to date format}
\usage{
fix_digit_date(date)
}
\arguments{
\item{date}{date as five number digit.}
}
\value{
Standardized date in a character format.
}
\description{
WIP: Convert dates that are stored as 5-digit number from Excel to date format
}
\examples{
fix_digit_date("44939") # res in "2023-01-13"

}
\seealso{
https://stackoverflow.com/a/50274507: depending on whether read in
file was created on Mac or Windows, the origin date might need to be adjusted.
}
