% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_fbg}
\alias{fix_fbg}
\title{Replace common textual descriptions with numeric values}
\usage{
fix_fbg(fbg)
}
\arguments{
\item{fbg}{character value.}
}
\value{
fbg transformed value.
}
\description{
Replace common textual descriptions with numeric values
}
