% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_gender}
\alias{fix_gender}
\title{Replace gender synonyms with either M, F or Other.}
\usage{
fix_gender(gender, id)
}
\arguments{
\item{gender}{patient gender.}

\item{id}{patient id.}
}
\value{
Either M, F or Other.
}
\description{
Replace gender synonyms with either M, F or Other.
}
