% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_id}
\alias{fix_id}
\title{check patient id and truncate it if necessary.}
\usage{
fix_id(id)
}
\arguments{
\item{id}{patient id.}
}
\value{
Either original id, truncated id, or error value.
}
\description{
check patient id and truncate it if necessary.
}
