% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_sex}
\alias{fix_sex}
\title{Replace sex synonyms with either M, F or Other.}
\usage{
fix_sex(sex, id)
}
\arguments{
\item{sex}{patient gender.}

\item{id}{patient id.}
}
\value{
Either M, F or Other.
}
\description{
Replace sex synonyms with either M, F or Other.
}
