% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_support_a4d}
\alias{fix_support_a4d}
\title{Fix invalid values for support a4d.}
\usage{
fix_support_a4d(support, id)
}
\arguments{
\item{support}{text}

\item{id}{patient id}
}
\value{
valid character value for support a4d.
}
\description{
Fix invalid values for support a4d.
}
