% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_testing_frequency}
\alias{fix_testing_frequency}
\title{Fix problems with ranges and decimal numbers in testing_frequency.}
\usage{
fix_testing_frequency(test_frq)
}
\arguments{
\item{test_frq}{character value for testing frequency}
}
\value{
validated test_frq value
}
\description{
Fix problems with ranges and decimal numbers in testing_frequency.
}
