% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{format_date_exceldate}
\alias{format_date_exceldate}
\title{format_date_exceldate}
\usage{
format_date_exceldate(product_df)
}
\description{
Reformates dates entered into excel in wrong format (e.g., dd-mm-yyyy) to final date format (yyyy-mm-dd)
}
