% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{format_date_excelnum}
\alias{format_date_excelnum}
\title{format_date_excelnum}
\usage{
format_date_excelnum(product_df)
}
\description{
Reformates dates entered into excel in wrong format (e.g., 44042) to final date format (yyyy-mm-dd)
}
