% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_main.R
\name{get_allowed_provinces}
\alias{get_allowed_provinces}
\title{Read allowed provinces from a YAML file.}
\usage{
get_allowed_provinces()
}
\value{
A named character vector with all allowed provinces.
}
\description{
Read in all provinces from a YAML file inside the provinces folder.
}
