% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_main.R
\name{get_files}
\alias{get_files}
\title{Find all files matching a search pattern in a given directory.}
\usage{
get_files(tracker_root, pattern = "\\\\.xlsx$")
}
\arguments{
\item{tracker_root}{The root directory to search in.}

\item{pattern}{The search pattern to filter files.}
}
\value{
A vector with file names.
}
\description{
Searches recursively for files matching a search pattern inside the root dir.
Only returns the file names without the paths.
}
