% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapping_table.R
\name{get_mapping_table}
\alias{get_mapping_table}
\title{Get mapping table for A4D}
\usage{
get_mapping_table(tracker_files, mapping_table_output)
}
\arguments{
\item{tracker_files}{The paths to the tracker files found using \code{get_tracker_files()}}

\item{mapping_table_output}{The output destination of the mapping table}
}
\value{
A mapping table containing the original and scrambled file names
and a .csv file that is exported to selected A4D directory
The scrambled names contain the year of the original file name and a hash
}
\description{
Get mapping table for A4D
}
