% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_main.R
\name{get_synonyms}
\alias{get_synonyms}
\title{Read synonyms from the synonyms YAML files.}
\usage{
get_synonyms()
}
\value{
A list with both patient and product data synonyms as tibble.
}
\description{
Read in all defined synonyms from the YAML files inside the synonyms folder.
}
