% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracker_year.R
\name{get_tracker_year}
\alias{get_tracker_year}
\title{Get the tracker year}
\usage{
get_tracker_year(tracker_data_file, month_list)
}
\arguments{
\item{tracker_data_file}{The path to the tracker file}

\item{month_list}{The list of all the monthly sheets within a tracker}
}
\value{
The year of the tracker
}
\description{
This function parses the tracker year from either the monthly sheets
or the tracker file name if the parsing is not successful
}
