% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{harmonize_input_data_columns}
\alias{harmonize_input_data_columns}
\title{harmonize_input_data_columns}
\usage{
harmonize_input_data_columns(product_df, columns_synonyms, sheet)
}
\arguments{
\item{product_df:}{The product data to be cleaned and matched against column synonyms}

\item{columns_synonyms:}{Long format output of read_column_synonyms to match columns}

\item{sheet:}{The current sheet name of the tracker file}
}
\description{
Imports the product df, cleans it and matches it against
}
