% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_product_data.R
\name{id_2_county_hospisal}
\alias{id_2_county_hospisal}
\title{Update country and hospital based on patient ID}
\usage{
id_2_county_hospisal(df, id, country, hospital)
}
\arguments{
\item{df}{A dataframe that contains the columns \code{ID}, \code{country}, and \code{hospital}.}

\item{id}{The name of the column in df that contains the patients ID information.}

\item{country}{The name of the column in df where the country information should be stored.}

\item{hospital}{The name of the column in df where the hospital information should be stored.}
}
\value{
The original dataframe with updated \code{country} and \code{hospital} columns.
}
\description{
This function updates the \code{country} and \code{hospital} columns in a dataframe based on the \code{ID} column.
It looks for rows where \code{ID} matches a specific pattern (two letters, an underscore, two more letters, and then digits).
For these rows, it puts the first two letters into \code{country} and the second two letters into \code{hospital}.
}
\examples{
df <- data.frame(
    id = c("US_CA123", "UK_LN456", "FR_PA789"),
    country = NA,
    hospital = NA,
    stringsAsFactors = FALSE
)
df <- id_2_product_county_hospisal(df, "ID", "country", "hospital")
}
