% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_main.R
\name{init_paths}
\alias{init_paths}
\title{initialize all necessary paths}
\usage{
init_paths(names, output_dir_name = "output", delete = FALSE)
}
\arguments{
\item{names}{Folder names that will be created under the output folder.}

\item{delete}{If TRUE, delete all files under output.}

\item{output_root_name}{The name of the main output folder created in the data folder.}
}
\value{
A list with tracker_root_path and output_root path
}
\description{
Create necessary output folder for main script under tracker_root_path.
This script creates a new output folder next to the data files, and
deletes all old files in it.
The script asks the user to set the tracker_root_path, either by
selecting a folder or reading from the A4D_DATA_ROOT env var.
}
