% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_product_patient.R
\name{link_product_patient}
\alias{link_product_patient}
\title{Link Product and Patient Data}
\usage{
link_product_patient(product_file, patient_file)
}
\arguments{
\item{product_file}{A string specifying the path to the product file.}

\item{patient_file}{A string specifying the path to the patient file.}
}
\value{
This function does not return a value. It prints log messages about the linking process.
}
\description{
This function links a product file with a patient file. It reads the files, merges them based on 'file_name' and patient 'id',
finds the ids that are in the product data but not in the patient data, creates a summary of these ids, and logs the process.
}
\examples{
\dontrun{
link_product_patient("path/to/product_data.parquet", "path/to/patient_data.parquet")
}
}
