% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{load_product_reference_data}
\alias{load_product_reference_data}
\title{Load Product List from Stock Summary}
\usage{
load_product_reference_data(
  stock_summary_xlsx = "master_tracker_variables.xlsx"
)
}
\arguments{
\item{stock_summary_xlsx}{A string that represents the path to the Excel file. Defaults to "master_tracker_variables.xlsx".}
}
\value{
A data frame that contains the product names. If there is an error during the process, it logs the error message.
}
\description{
This function loads the product list from 'Stock_Summary' sheet in an Excel file.
}
\examples{
\dontrun {
product_list <- load_product_reference_data()
product_list <- load_product_reference_data("your_file.xlsx")
}
}
