% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_to_json}
\alias{log_to_json}
\title{Log-To-Json Formatter}
\usage{
log_to_json(
  message,
  values = NA,
  script = NA,
  file = NA,
  errorCode = NA,
  warningCode = NA,
  functionName = NA
)
}
\arguments{
\item{message}{log message with glue syntax {values\link{'X'}}.}

\item{values}{A named list with the glue referenced values as single strings.}

\item{script}{script name where the log message was created.}

\item{file}{file name where the log message was created.}

\item{errorCode}{A unique error code for easy aggregation of error messages.}

\item{warningCode}{A unique warning code for easy aggregation of warning messages.}

\item{functionName}{The name of the function where the log message was created.}
}
\value{
stringified JSON to log with logX functions
}
\description{
Log-To-Json Formatter
}
