% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{parse_dates}
\alias{parse_dates}
\title{Try to handle invalid dates.}
\usage{
parse_dates(date)
}
\arguments{
\item{date}{character date value.}
}
\value{
Either a correctly parsed date or NA.
}
\description{
For each value of the character vector, this function uses the
lubridate parse functions to find common formats
}
