% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_dates.R
\name{parse_invalid_dates}
\alias{parse_invalid_dates}
\title{Try to handle invalid dates.}
\usage{
parse_invalid_dates(df, colname)
}
\arguments{
\item{df}{data frame with patient data.}

\item{colname}{column that has invalid dates.}
}
\description{
For each value of the character vector, this function uses the
lubridate parse functions to find common formats
}
