% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_product_data.R
\name{preparing_product_fields}
\alias{preparing_product_fields}
\title{Preparing Product Fields}
\usage{
preparing_product_fields(merged_data)
}
\arguments{
\item{merged_data}{A data frame that needs to be processed.}
}
\value{
A data frame with processed fields.
}
\description{
This function processes fields for a single csv product_data.
It checks if all fields are present in merged_data and adds missing fields with NA.
It ensures the correct data type for each column and reorders the columns according to the list of fields.
Any additional fields in the input dataframe are moved to the end.
}
\examples{
\dontrun{
df <- data.frame(product = c("A", "B"), product_units_notes = c("note1", "note2"))
df <- preparing_product_fields(df)
}

}
