% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_script_1.R
\name{process_patient_data}
\alias{process_patient_data}
\title{Extract patient data.}
\usage{
process_patient_data(
  tracker_name,
  tracker_data_file,
  output_root,
  synonyms_patient
)
}
\arguments{
\item{tracker_name}{Filename without extension.}

\item{tracker_data_file}{Filename of the tracker.}

\item{output_root}{Directory for storing extracted patient data.}

\item{synonyms_patient}{Synonyms for patient data header names.}
}
\description{
Extract patient data.
}
