% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_script_2.R
\name{process_patient_file}
\alias{process_patient_file}
\title{Process and clean patient raw data}
\usage{
process_patient_file(patient_file, paths, allowed_provinces)
}
\arguments{
\item{patient_file}{Path to patient raw parquet file from first script.}

\item{paths}{List with paths to patient_data_cleaned and product_data_cleaned folders.}

\item{allowed_provinces}{A named character vector with names of allowed provinces.}
}
\description{
Read in the output of the first script as parquet (all character).
Merge extracted patient data with the metadata schema that contains all
necessary columns. Process each variable in three steps:
\enumerate{
\item Fix any problems with character representation.
\item Convert column to target data type, for example a date or number.
\item Apply any post-processing logic for the target data type.
}
}
