% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_script_1.R
\name{process_product_data}
\alias{process_product_data}
\title{Extract product data.}
\usage{
process_product_data(
  tracker_name,
  tracker_data_file,
  output_root,
  synonyms_product
)
}
\arguments{
\item{tracker_name}{Filename without extension.}

\item{tracker_data_file}{Filename of the tracker.}

\item{output_root}{Directory for storing extracted product data.}

\item{synonyms_product}{Synonyms for product data header names.}
}
\description{
Extract product data.
}
