% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script2_process_patient_data.R
\name{process_raw_patient_file}
\alias{process_raw_patient_file}
\title{Process raw patient data}
\usage{
process_raw_patient_file(paths, patient_file, patient_file_name, output_root)
}
\arguments{
\item{paths}{list of paths to the input and output directories.}

\item{patient_file}{name of the parquet file with the raw patient data.}

\item{patient_file_name}{base name of the parquet file with the raw patient data.}

\item{output_root}{root directory of the output folder.}
}
\description{
This function reads the raw patient data from a parquet file, cleans the data,
and exports the cleaned data as a parquet file.
}
