% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script1_process_tracker_file.R
\name{process_tracker_file}
\alias{process_tracker_file}
\title{Process tracker file.}
\usage{
process_tracker_file(paths, tracker_file, tracker_name, synonyms)
}
\arguments{
\item{paths}{list of paths to directories.}

\item{tracker_file}{Name of the tracker file.}

\item{tracker_name}{Basename of the tracker file.}

\item{synonyms}{list of synonyms for patient and product data.}
}
\description{
Process tracker file.
}
