% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script1_process_patient_data.R
\name{process_tracker_patient_data}
\alias{process_tracker_patient_data}
\title{Process tracker patient data}
\usage{
process_tracker_patient_data(
  tracker_name,
  tracker_data_file,
  output_root,
  synonyms_patient
)
}
\arguments{
\item{tracker_name}{The name of the tracker.}

\item{tracker_data_file}{The path to the tracker data file.}

\item{output_root}{The root directory of the output folder.}

\item{synonyms_patient}{A named list of synonyms for the patient data columns.}
}
\description{
This function reads the patient data from a CSV file, adds a file name column,
and exports the data as a parquet file.
}
