% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script1_process_product_data.R
\name{process_tracker_product_data}
\alias{process_tracker_product_data}
\title{Process tracker product data}
\usage{
process_tracker_product_data(
  tracker_name,
  tracker_data_file,
  output_root,
  synonyms_product
)
}
\arguments{
\item{tracker_name}{The name of the tracker.}

\item{tracker_data_file}{The path to the tracker data file.}

\item{output_root}{The root directory of the output folder.}

\item{synonyms_product}{A named list of synonyms for the product data columns.}
}
\description{
This function reads the product data from an Excel file,
fills missing values, and exports the data as a parquet file.
}
