% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cleaned_patient_data.R
\name{read_cleaned_patient_data}
\alias{read_cleaned_patient_data}
\title{Read in all processed and cleaned patient data files}
\usage{
read_cleaned_patient_data(input_root, patient_data_files)
}
\arguments{
\item{input_root}{root directory containing all files.}

\item{patient_data_files}{list of CSV files with cleaned patient data.}
}
\value{
data frame holding all cleaned patient data with correct column types.
}
\description{
Read in all processed and cleaned patient data files
}
