% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_main.R
\name{read_column_synonyms}
\alias{read_column_synonyms}
\title{Get all synonyms for all variable names}
\usage{
read_column_synonyms(synonym_file)
}
\arguments{
\item{synonym_file}{A YAML file containing the synonyms}
}
\value{
A tibble containing unique column names and their synonyms.
}
\description{
This function reads the synonyms from a YAML file
and generates a tibble containing unique column names and their synonyms.
}
\examples{
\dontrun{
read_column_synonyms(synonym_file = "synonyms_patient.yaml")
read_column_synonyms(synonym_file = "synonyms_product.yaml")
}
}
