% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{read_column_synonyms_product}
\alias{read_column_synonyms_product}
\title{read_column_synonyms_product}
\usage{
read_column_synonyms_product(codebook_data_file)
}
\description{
Imports the codebook, cleans, removes duplicates and transforms it
into long df format
}
