% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_patient_an_data.R
\name{read_patient_an_data}
\alias{read_patient_an_data}
\title{Get patient data from "Patient AN Data" sheet}
\usage{
read_patient_an_data(tracker_data_file, sheet_list, columns_synonyms)
}
\arguments{
\item{tracker_data_file}{description}

\item{sheet_list}{description}

\item{columns_synonyms}{description}
}
\value{
description
}
\description{
Extract the data in the "Patient AN Data" sheet
as data.frame.
The data is cleaned and the column headers are sanitized and
replaced with their standard variable name.
}
