% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_main.R
\name{read_patient_csv}
\alias{read_patient_csv}
\title{Read in patient data from CSV.}
\usage{
read_patient_csv(patient_file_path)
}
\arguments{
\item{patient_file_path}{Path to the CSV file.}
}
\value{
tibble with patient data
}
\description{
Read in patient data from CSV.
}
