% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script1_read_patient_data.R
\name{reading_patient_data}
\alias{reading_patient_data}
\title{Reading patient data from the tracker data file}
\usage{
reading_patient_data(tracker_data_file, columns_synonyms)
}
\arguments{
\item{tracker_data_file}{The path to the tracker data file.}

\item{columns_synonyms}{A named list of column synonyms.}
}
\value{
A data frame with the patient data.
}
\description{
This function reads the patient data from the tracker data file,
harmonizes the columns, and returns the data as a data frame.
}
