% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{reading_product_data_step2}
\alias{reading_product_data_step2}
\title{Process product data in script 2.}
\usage{
reading_product_data_step2(df, columns_synonyms)
}
\arguments{
\item{df}{Dataframe. Output of tracker file from script 1 for product data.}

\item{columns_synonyms}{Tibble with synonyms for product data (columns: name_clean and name_to_be_matched)}
}
\value{
Cleaned product data for one specified tracker.
}
\description{
Processes data output from script 1. Data will be further cleaned and processed to fix data quality issues and to create a report.
}
