% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_format_2_long_format.R
\name{remove_original_cells}
\alias{remove_original_cells}
\title{Remove Original Cells based on Relationship}
\usage{
remove_original_cells(df, col_units_rel, col_rel_to)
}
\arguments{
\item{df}{A data frame containing the columns \code{col_units_rel} and \code{col_rel_to}.}

\item{col_units_rel}{The name of the column containing units (cells) potentially related to multiple other units.}

\item{col_rel_to}{The name of the column containing relationships between units, where comma (",") indicates a relationship with multiple units.}
}
\value{
The modified data frame where rows with comma-separated relationships in \code{col_rel_to} have corresponding values in \code{col_units_rel} set to NA and the comma-separated values removed from \code{col_rel_to}.
}
\description{
This function removes rows from the column \code{col_units_rel} where the corresponding value in \code{col_rel_to} contains a comma (",").
It also removes those comma-separated values from the \code{col_rel_to} column itself.
}
\examples{
# Create a sample data frame
df <- data.frame(col_units_rel = c(1, 2, 3, 4), col_rel_to = c("A", "B,C", "D", "A,B"))

# Remove original cells based on relationship
df_modified <- remove_original_cells(df, col_units_rel = "col_units_rel", col_rel_to = "col_rel_to")

# Print the modified data frame
print(df_modified)
}
