% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{remove_rows_with_na_columns}
\alias{remove_rows_with_na_columns}
\title{Remove Rows with NA Values in Specified Columns.}
\usage{
remove_rows_with_na_columns(df, column_names)
}
\arguments{
\item{df}{A data frame that contains the data from which rows need to be removed.}

\item{column_names}{A character vector specifying the column names to be checked for NA values.}
}
\value{
A new data frame with rows removed if all the specified columns in the row contain NA values. The returned data frame will have the same structure as the input data frame, but with rows that satisfy the condition removed.
}
\description{
This function takes a data frame and a vector of column names as input. It removes rows from the data frame where all the specified columns contain only NA values. If any of the specified columns have at least one non-NA value, the corresponding row will be retained in the output data frame.
}
