% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_product_data.R
\name{reorder_product_fields}
\alias{reorder_product_fields}
\title{Reorder and Add Missing Fields in a Dataframe}
\usage{
reorder_product_fields(merged_data)
}
\arguments{
\item{merged_data}{A dataframe that needs its columns to be reordered and missing fields added.}
}
\value{
A dataframe with reordered columns and added missing fields.
}
\description{
This function checks if all specified fields are present in the input dataframe.
If any fields are missing, it adds them with NA values.
It then reorders the columns according to the specified list of fields.
Any additional fields in the input dataframe are moved to the end.
}
\examples{
\dontrun{
df <- data.frame(product = c("A", "B"), product_units_notes = c("note1", "note2"))
df <- reorder_product_fields(df)
}

}
