% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{replace_empty_string_with_NA}
\alias{replace_empty_string_with_NA}
\title{replace empty strings with NA.}
\usage{
replace_empty_string_with_NA(string_vector)
}
\arguments{
\item{string_vector}{character vector containing empty strings.}
}
\description{
replace empty strings with NA.
}
