% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{replace_range_with_mean}
\alias{replace_range_with_mean}
\title{Replace a range with mean.}
\usage{
replace_range_with_mean(x)
}
\arguments{
\item{x}{range as text..}
}
\value{
mean of that range
}
\description{
A range like "0-2" will be replaced with the mean(c(0,2)).
}
