% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_product_data.R
\name{report_empty_intersections}
\alias{report_empty_intersections}
\title{Report Empty Intersections}
\usage{
report_empty_intersections(df, row_category, col_category)
}
\arguments{
\item{df}{A dataframe.}

\item{row_category}{The name of the column in df to use for the rows of the cross-tabulation.}

\item{col_category}{The name of the column in df to use for the columns of the cross-tabulation.}
}
\value{
A warning message with the number of row_category values that have a sum of 0 in col_category, and a list of these row_category values.
}
\description{
This function reports the names of the rows in a cross-tabulation of two columns
from a dataframe where the sum is 0.
}
\examples{
df <- data.frame(file_name = c("file1", "file2", "file3"), table_country = c("country1", "country2", NA))
report_empty_intersections(df, "file_name", "table_country")
}
