% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_clean_data.R
\name{sanitize_column_name}
\alias{sanitize_column_name}
\title{Sanitize column name}
\usage{
sanitize_column_name(column_name)
}
\arguments{
\item{column_name}{A string representing the name of the column to be sanitized.}
}
\value{
The sanitized column name.
}
\description{
This function takes in a column name,
sanitizes it by converting all letters to lower case,
removing all spaces and special characters except alphanumeric characters,
and then returns the sanitized column name.
}
\examples{
sanitize_column_name("John Doe's Column") # should return "johndoescolumn"
sanitize_column_name("Date 2022")         # should return "date2022"
sanitize_column_name("My Awesome 1st Column!!") # should return "myawesome1stcolumn"

}
