% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_a4d_data_root.R
\name{set_a4d_data_root}
\alias{set_a4d_data_root}
\title{Helper function that sets the env variable to the A4D tracker files}
\usage{
set_a4d_data_root()
}
\value{
Returns a character representing the path to the tracker data
}
\description{
Helper function that sets the env variable to the A4D tracker files
}
