% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{set_second_row_as_headers_and_remove_first_row}
\alias{set_second_row_as_headers_and_remove_first_row}
\title{set_second_row_as_headers_and_remove_first_row}
\usage{
set_second_row_as_headers_and_remove_first_row(df)
}
\description{
Clean df where first row is empty
}
