% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{setup_logger}
\alias{setup_logger}
\title{Configure basic logger}
\usage{
setup_logger(output_dir, log_name)
}
\arguments{
\item{output_dir:}{Output directory of the main script}
}
\description{
Configure basic logger
}
