% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{split_bp_in_sys_and_dias}
\alias{split_bp_in_sys_and_dias}
\title{Separate single blood pressure value into sys and dias values.}
\usage{
split_bp_in_sys_and_dias(df)
}
\arguments{
\item{df}{data frame with patient data.}
}
\value{
data frame with two new columns: blood_pressure_sys_mmhg and blood_pressure_dias_mmhg.
}
\description{
Separate single blood pressure value into sys and dias values.
}
