% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_product_data.R
\name{switch_columns_stock}
\alias{switch_columns_stock}
\title{Switch product_received_from and product_units_received column.}
\usage{
switch_columns_stock(df)
}
\arguments{
\item{df}{Dataframe. Output of tracker file from script 1 for product data.}
}
\value{
Dataframe with columns switched (renamed)
}
\description{
Renaming columns when string Remaining stock in wrong column. E.g. 2018_PNG for Nov and Dec.
}
