% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_product_data.R
\name{update_receivedfrom}
\alias{update_receivedfrom}
\title{Transfers in trackers with specific format which have "START BALANCE" in product data the start balance to the correct column.}
\usage{
update_receivedfrom(product_df)
}
\arguments{
\item{product_df}{Input product data.}
}
\value{
Dataframe with corrected product_received_from column extracting the starting balance from trackers where necessary.
}
\description{
Run before clean_receivedfrom and recode_unitcolumnstozero. Necessary for specific format such as in 2019_PKH and 2020_STH. When start balance in tracker next to text "start balance" one need to extract the balance either from the "received from" column or the "units released" column.
}
