% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{with_file_logger}
\alias{with_file_logger}
\title{Temporary File Logger}
\usage{
with_file_logger(new, code, output_root)
}
\arguments{
\item{new}{New filename for the logfile}

\item{code}{Code to be executed}

\item{output_root}{Root output folder from where to construct the log path}
}
\description{
Temporarily change logging to exclusively log to specific log file
}
