% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{check_allowed_values}
\alias{check_allowed_values}
\title{Check value against list of allowed values.}
\usage{
check_allowed_values(x, valid_values, error_val, id, col = "")
}
\arguments{
\item{x}{character value to check.}

\item{valid_values}{list of valid options.}

\item{error_val}{value that is used to signal invalid values.}

\item{id}{patient id.}

\item{col}{column name when used with mutate/across.}
}
\value{
Either x or "Other".
}
\description{
Check value against list of allowed values.
}
