% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{convert_to}
\alias{convert_to}
\title{Try to convert x into a target type.}
\usage{
convert_to(x, cast_fnc, error_val)
}
\arguments{
\item{x}{value to convert to numeric}

\item{cast_fnc}{function to use to cast x}

\item{error_val}{error value to use when as.numeric fails}
}
\value{
converted x
}
\description{
Use cast_fnc to cast x to target type.
If that cast fails, output error_val instead.
If x is already NA, output NA
}
