% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_patient_data_changes_only.R
\name{create_table_longitudinal_data}
\alias{create_table_longitudinal_data}
\title{Create CSV with longitudinal patient data for a single variable.}
\usage{
create_table_longitudinal_data(
  patient_data_files,
  input_root,
  output_root,
  variable,
  name
)
}
\arguments{
\item{patient_data_files}{list of CSV files with cleaned patient data from step 2.}

\item{input_root}{root directory of the input CSV files.}

\item{output_root}{root directory of the output folder.}

\item{variable}{name of the column that should be exported.}

\item{name}{name used to create the export file name.}
}
\description{
Read in all cleaned patient data CSV and create a single data.frame.
Group this data by id and take only the months when there is a change in the medical data.
}
