% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_patient_data.R
\name{create_table_patient_data_monthly}
\alias{create_table_patient_data_monthly}
\title{Create CSV with monthly patient data}
\usage{
create_table_patient_data_monthly(patient_data_files, input_root, output_root)
}
\arguments{
\item{patient_data_files}{list of CSV files with cleaned patient data from step 2.}

\item{input_root}{root directory of the input CSV files.}

\item{output_root}{root directory of the output folder.}
}
\description{
Read in all cleaned patient data CSV and create a single data.frame.
Only take dynamic columns.
}
