% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_product_data.R
\name{create_table_product_data}
\alias{create_table_product_data}
\title{Create a Single CSV for Table Product Data}
\usage{
create_table_product_data(input_root, output_root)
}
\arguments{
\item{input_root}{A string specifying the path to the directory containing the input CSV files.}

\item{output_root}{A string specifying the path to the directory where the output CSV file should be written.}
}
\value{
This function does not return a value. It writes the merged data to a new CSV file
(with reordered columns according to the list of fields) in the output_root directory.
}
\description{
This function reads all CSV files in a specified directory, merges them into a single data frame,
and writes the merged data to a new CSV file with reordered and processed columns according to the list of fields.
If there are any columns that are present in some CSV files but not others,
this function will add those missing columns to the data frames where they are missing and fill them with NA.
}
\examples{
\dontrun{
create_table_product_data("path/to/input/directory", "path/to/output/directory")
}
}
