% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_bmi}
\alias{fix_bmi}
\title{Calculate bmi value based on weight and height.}
\usage{
fix_bmi(weight, height, id)
}
\arguments{
\item{weight}{numeric weight value.}

\item{height}{numeric height value.}

\item{id}{patient id.}
}
\value{
calculated bmi value.
}
\description{
Formula: weight (kg) / height^2 (meters^2)
bmi is set to NA if either weight or height are NA.
bmi is set to error val if either weight or height are error val.
}
