% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_patient_data_fix.R
\name{fix_t1d_diagnosis_age}
\alias{fix_t1d_diagnosis_age}
\title{Try to replace text descriptions for age with proper numerical values.}
\usage{
fix_t1d_diagnosis_age(t1d_diagnosis_age, t1d_diagnosis_date, id)
}
\arguments{
\item{t1d_diagnosis_age}{patient age when diagnosed with T1D.}

\item{t1d_diagnosis_date}{date when diagnosed with T1D.}

\item{id}{patient id}
}
\description{
Try to replace text descriptions for age with proper numerical values.
}
