% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_to_json}
\alias{log_to_json}
\title{Log-To-Json Formatter}
\usage{
log_to_json(
  message,
  values = NA,
  file = NA,
  errorCode = NA,
  warningCode = NA,
  functionName = NA
)
}
\arguments{
\item{message}{Log Message with Glue Syntax {values\link{'X'}}}

\item{values}{A named list with the Glue Referenced Values as single Strings}

\item{errorCode}{A Combination of UID Letters for Easy location of the Message Origin}

\item{warningCode}{A Combination of UID Letters for Easy location of the Message Origin}

\item{currentFile}{String of current File with Path the Log is evoked}
}
\value{
Stringified JSON to Log with logX-Functions
}
\description{
Log-To-Json Formatter
}
