% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_script_2.R
\name{process_product_file}
\alias{process_product_file}
\title{Process and clean product raw data}
\usage{
process_product_file(product_file, paths, p)
}
\arguments{
\item{product_file}{Path to product raw parquet file from first script.}

\item{paths}{List with paths to patient_data_cleaned and product_data_cleaned folders.}

\item{p}{progressor from progressr package.}
}
\description{
A short description...
}
