% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scipt_1.R
\name{process_tracker_file}
\alias{process_tracker_file}
\title{Process a single tracker file and extract patient and product data.}
\usage{
process_tracker_file(tracker_file, paths, p)
}
\arguments{
\item{tracker_file}{Filename of the tracler.}

\item{paths}{a list with the paths to the tracker root dir, the patient and product output dir and the root output dir.}

\item{p}{progressor from progressr package.}
}
\description{
Process a single tracker file and extract patient and product data.
}
