% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_column_synonyms.R
\name{read_column_synonyms}
\alias{read_column_synonyms}
\title{Get all synonyms for all variable names}
\usage{
read_column_synonyms(master_tracker_file, sheet)
}
\arguments{
\item{master_tracker_file}{An excel file containing the synonyms.}

\item{sheet}{The name of the sheet to read from the excel file.}
}
\value{
A tibble containing unique column names and their synonyms.
}
\description{
This function reads a specific sheet from an excel file
and generates a tibble containing unique column names and their synonyms.
}
\examples{
\dontrun{
read_column_synonyms("master_tracker_variables.xlsx", "synonyms_PatientData")
}
}
