% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_A4D_directory.R
\name{select_A4D_directory}
\alias{select_A4D_directory}
\title{Selects path to A4D data and sets it as an environment variable}
\usage{
select_A4D_directory()
}
\value{
Sets the env variable to A4D data folder
}
\description{
Selects path to A4D data and sets it as an environment variable
}
