% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{setup_file_logger}
\alias{setup_file_logger}
\title{Add a basic default file logger for a newly processed file.}
\usage{
setup_file_logger(output_root, logfile)
}
\arguments{
\item{output_root}{Output root directory for the current process.}

\item{logfile}{The name of the log file.}
}
\description{
Add a basic default file logger for a newly processed file.
}
